+++ packages/test-renderer/src/index.tsx
@@ -19,7 +19,18 @@ extend(THREE as any)
 const create = async (element: React.ReactNode, options?: Partial<CreateOptions>): Promise<Renderer> => {
   const canvas = createCanvas(options)
 
-  const _root = createRoot(canvas).configure({ frameloop: 'never', ...options, events: undefined })
+  const _root = createRoot(canvas).configure({
+    frameloop: 'never',
+    // TODO: remove and use default behavior
+    size: {
+      width: options?.width ?? 1280,
+      height: options?.height ?? 800,
+      top: 0,
+      left: 0,
+    },
+    ...options,
+    events: undefined,
+  })
   const _store = mockRoots.get(canvas)!.store
 
   await act(async () => _root.render(element))